Response handling: Reproduce wrong response sent in some cases
Note: This PR is just an observation, as this repo is still a prototype. It helps reproducing a small
issue, but it is not meant to be merged.

When the endpoint does not use Send() to send the response, RadEndpoint still sends the response
back, but a different instance of it, rather than the expected one. This causes Response to differ
compared to what is done with it in the endpoint.

 - The issue of not using Send() is easy to make.

 - There are no compile time, or run time checks if it is not invoked.

 - Potential solutions:
   1. Require return value in the Handle method enforcing returning a response.
   2. Require response as a parameter of Send and not have Response as a premade instance.
     to make clar that Response somehow has to be sent back.
   3. Combine point #2 with an analyzer to provide compile time hint.